// frontend/src/client/models/ExistingCustomGPT.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExistingCustomGPT = {
    custom_gpt_id: number;
    custom_gpt_name: string;
    custom_gpt_description: string;
    custom_gpt_instructions: string;
    created_at: (string | null);
};